services/backend-fastapi/pyproject.toml
[project]
name = "charlott-fastapi"
version = "0.1.0"
description = "Charlott CRM FastAPI backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.36",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.13.2",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "redis>=5.0.8",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.23.8",
  "ruff>=0.6.8",
  "mypy>=1.11.2",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
